Adjusting the axis range
I made so that the x axis and the y axis have the same range.

With the default setting. there seems to be a more pronounced difference in the y axis than there is in fact because the y-axis had a very narrow range.

Increase font size for the plot